kernel seek implementation

+++ kernel/include/kernel/fs/vfs.h
@@ -16,7 +16,7 @@ struct file *fs_open(const char *file_name, int *error);
 int fs_close(struct file *file);
 ssize_t fs_read(struct file *file, void *buffer, size_t len);
 ssize_t fs_write(struct file *file, const void *buffer, size_t len);
-int fs_seek(struct file *file, off_t offset, int whence);
+off_t fs_seek(struct file *file, off_t offset, int whence);
 long fs_tell(struct file *file);
 int fs_stat(const char *file_name, struct stat *stat_struct);
 int fs_mount(const char *src, const char *path, const char *type);
